move some code to its own file
